Test commit

+++ src/main/java/parking/guru/config/security/UserDetailsServiceImpl.java
@@ -20,7 +20,6 @@ public class UserDetailsServiceImpl implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String username) {
-        // Use getUserByEmail instead of getUserByUserEmail
         User user = userService.getUserByEmail(username)
                 .orElseThrow(() -> new UsernameNotFoundException(String.format("Username %s not found", username)));
         List<SimpleGrantedAuthority> authorities = Collections.singletonList(new SimpleGrantedAuthority(String.valueOf(user.getRole())));
@@ -29,7 +28,7 @@ public class UserDetailsServiceImpl implements UserDetailsService {
 
     private CustomUserDetails mapUserToCustomUserDetails(User user, List<SimpleGrantedAuthority> authorities) {
         CustomUserDetails customUserDetails = new CustomUserDetails();
-        customUserDetails.setId(user.getId());
+        customUserDetails.setId((long) user.getId());
         customUserDetails.setPassword(user.getPassword());
         customUserDetails.setEmail(user.getEmail());
         customUserDetails.setAuthorities(authorities);
